coingecko/requirements.txt
arrow==1.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
progressbar2==4.4.2
pycoingecko==3.1.0
python-dateutil==2.9.0.post0
python-utils==3.8.2
requests==2.31.0
six==1.16.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
urllib3==2.2.1
